+++ tools/decode-status.py
@@ -201,7 +201,8 @@ a_setoption = [[
     "(Zigbee) Include time in `ZbReceived` messages like other sensors",
     "(MQTT) Retain on Status"
     ],[
-    "","","","",
+    "(ESP32) Show ESP32 internal temperature sensor",
+    "","","",
     "","","","",
     "","","","",
     "","","","",
@@ -319,7 +320,7 @@ else:
         obj = json.load(fp)
 
 def StartDecode():
-    print ("\n*** decode-status.py v12.1.0.1 by Theo Arends and Jacek Ziolkowski ***")
+    print ("\n*** decode-status.py v12.1.1.1 by Theo Arends and Jacek Ziolkowski ***")
 
 #    print("Decoding\n{}".format(obj))